Swapped trenary operator to if/else
This way we can verify the existence of data[i].newValue without having to deal with the trenary operator type restrictions. (in this cause it caused a conflict between char* and wchar_t*).

Having only tested on BnS it worked fine but would have caused crash for GW2 auto generated configs as pointed by megai2.

Co-Authored-By: megai2 <megai2@ya.ru>